Remove dh dir, move all SDK files to parent dir
