+++ info.go
@@ -1,4 +1,4 @@
-package dh
+package devicehive_go
 
 type ServerInfo struct {
 	APIVersion         string      `json:"apiVersion"`
