+++ transportadapter/utils.go
